Refactor should eager load
